requirements.txt
# Web servers
fastapi
uvicorn[standard]

# Frontend (Streamlit)
streamlit

# File handling
python-multipart
aiofiles
pydub

# Scoring / logic
requests
jiwer
transformers
torchaudio
soundfile

# TTS
gTTS

# ASR
whisper
faster-whisper
# Environment variables
python-dotenv


jiwer
transformers
torchaudio
pydub
soundfile

#Database
sqlalchemy 
psycopg2-binary 
alembic